feat: Simplistic entity registry and initialization process (#15)

+++ src/core/components/TransformComponent.hh
@@ -12,12 +12,12 @@ class TransformComponent : public AbstractComponent
   TransformComponent(IBoundingBoxShPtr bbox);
   ~TransformComponent() override = default;
 
-  auto position() const -> Eigen::Vector3f;
+  auto position() const -> Eigen::Vector3d;
   auto size() const -> float;
-  bool contains(const Eigen::Vector3f &pos) const;
+  bool contains(const Eigen::Vector3d &pos) const;
 
-  void translate(const Eigen::Vector3f &delta);
-  void overridePosition(const Eigen::Vector3f &position);
+  void translate(const Eigen::Vector3d &delta);
+  void overridePosition(const Eigen::Vector3d &position);
 
   void simulate(const time::TickData &data) override;
 
@@ -27,7 +27,7 @@ class TransformComponent : public AbstractComponent
   /// @param localPos - the position to transform, expressed in the local coordinate
   /// frame
   /// @return - the position expressed in the global coordinate frame
-  auto transformToGlobal(const Eigen::Vector3f &localPos) const -> Eigen::Vector3f;
+  auto transformToGlobal(const Eigen::Vector3d &localPos) const -> Eigen::Vector3d;
 
   private:
   IBoundingBoxShPtr m_bbox{};
