+++ src/core/physics/bbox/CircleBox.cc
@@ -4,33 +4,33 @@
 
 namespace swarms::core {
 
-CircleBox::CircleBox(const Eigen::Vector3f &center, const float radius)
+CircleBox::CircleBox(const Eigen::Vector3d &center, const double radius)
   : m_center(center)
   , m_radius(radius)
   , m_squaredRadius(m_radius * m_radius)
 {}
 
-auto CircleBox::radius() const -> float
+auto CircleBox::radius() const -> double
 {
   return m_radius;
 }
 
-auto CircleBox::position() const -> Eigen::Vector3f
+auto CircleBox::position() const -> Eigen::Vector3d
 {
   return m_center;
 }
 
-void CircleBox::moveTo(const Eigen::Vector3f &position)
+void CircleBox::moveTo(const Eigen::Vector3d &position)
 {
   m_center = position;
 }
 
-bool CircleBox::isInside(const Eigen::Vector3f &pos) const
+bool CircleBox::isInside(const Eigen::Vector3d &pos) const
 {
   return (pos - m_center).squaredNorm() < m_squaredRadius;
 }
 
-void CircleBox::translate(const Eigen::Vector3f &delta)
+void CircleBox::translate(const Eigen::Vector3d &delta)
 {
   m_center += delta;
 }
